requirements.txt
-e .
pytest

aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.aws_lambda_python
aws-cdk.core
python-dotenv